Change case when amto.conf is already present 	modified:   11_apache_configuration.sh

+++ src/11_apache_configuration.sh
@@ -36,9 +36,11 @@ main(){
   #Download appropriate modules, copy the dist, add the macro
   a2enmod rewrite ssl proxy proxy_http auth_cas wsgi macro headers 2>> $output_for_logs
   /etc/init.d/apache2 restart
-  cp "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf.dist" "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf" 2>> $output_for_logs
-  echo "Use Atmosphere $SERVERNAME $SERVERNAME.iplantc.org $LOCATIONOFATMOSPHERE $LOCATIONOFTROPOSPHERE $NAMEOFYOURSSLCERTIFICATE $NAMEOFYOURSSLKEY $NAMEOFYOURSSLBUNDLECERTIFICATE" >> "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf" 2>> $output_for_logs
 
+  if [ ! -e "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf" ]; then
+    cp "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf.dist" "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf" 2>> $output_for_logs
+    echo "Use Atmosphere $SERVERNAME $SERVERNAME.iplantc.org $LOCATIONOFATMOSPHERE $LOCATIONOFTROPOSPHERE $NAMEOFYOURSSLCERTIFICATE $NAMEOFYOURSSLKEY $NAMEOFYOURSSLBUNDLECERTIFICATE" >> "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf" 2>> $output_for_logs
+  fi
   #Symlink all extras files to appropriate space in apache so that future
   #updates are applied automatically
   ln -s "$LOCATIONOFATMOSPHERE/extras/apache/atmo.conf" /etc/apache2/sites-available/atmo.conf 2>> $output_for_logs
